python/requirements_new.txt
ImageHash==4.3.2
numpy==2.4.3
opencv-python==4.13.0.92
pillow==12.1.1
PyWavelets==1.9.0
scipy==1.17.1
google-cloud-pubsub==2.19.0
firebase-admin==6.4.0
google-generativeai==0.6.0
mediapipe>=0.10.30
deepface>=0.0.79
tf-keras>=2.15.0
requests==2.32.3
google-api-python-client>=2.108.0
praw>=7.7.0
